requirements.txt
fastapi
uvicorn[standard]
chromadb
transformers
torch
google-generativeai
pydantic
requests
numpy
